Refactor confusing naming of tests
