+++ test/integration_tests/integrationtest_ros_scanner_node.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2019-2020 Pilz GmbH & Co. KG
+// Copyright (c) 2020 Pilz GmbH & Co. KG
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Lesser General Public License as published by
@@ -13,47 +13,121 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-#include <string>
 #include <memory>
+#include <chrono>
+#include <thread>
+#include <future>
 
 #include <gtest/gtest.h>
+#include <gmock/gmock.h>
+
+#include <ros/ros.h>
+
+#include <sensor_msgs/LaserScan.h>
+#include <pilz_utils/wait_for_message.h>
+#include <pilz_testutils/async_test.h>
 
 #include "psen_scan/ros_scanner_node.h"
-#include "psen_scan/build_ros_message_exception.h"
+#include "psen_scan/laserscan.h"
 #include "psen_scan/mock_scanner.h"
+#include "psen_scan/psen_scan_fatal_exception.h"
 
 using namespace psen_scan;
+using namespace psen_scan_test;
+using ::testing::DoAll;
+using ::testing::Return;
 
 namespace psen_scan_test
 {
-TEST(BuildRosMsgFromLaserScanTests, buildROSMessage)
+static constexpr std::chrono::seconds LOOP_END_TIMEOUT{ 3 };
+
+static constexpr int QUEUE_SIZE{ 10 };
+
+static const std::string LASER_SCAN_RECEIVED{ "LASER_SCAN_RECEIVED" };
+
+class SubscriberMock
 {
-  ros::NodeHandle nh;
-  std::unique_ptr<MockScanner> mock_scanner{ new MockScanner() };
-  ROSScannerNode ros_scanner_node(nh, "node1_topic", "node1_frame", Degree(137.5), std::move(mock_scanner));
+public:
+  void initialize(ros::NodeHandle& nh);
 
-  LaserScan laser_scan_fake{ PSENscanInternalAngle(1), PSENscanInternalAngle(1), PSENscanInternalAngle(2) };
+  MOCK_METHOD1(callback, void(const sensor_msgs::LaserScan& msg));
+
+private:
+  ros::Subscriber subscriber_;
+};
+
+inline void SubscriberMock::initialize(ros::NodeHandle& nh)
+{
+  subscriber_ = nh.subscribe("scan", QUEUE_SIZE, &SubscriberMock::callback, this);
+}
+
+class RosScannerNodeTests : public testing::Test, public testing::AsyncTest
+{
+protected:
+  ros::NodeHandle nh_priv_{ "~" };
+};
+
+TEST_F(RosScannerNodeTests, testScanTopicReceived)
+{
+  SubscriberMock subscriber;
+  EXPECT_CALL(subscriber, callback(::testing::_)).WillOnce(ACTION_OPEN_BARRIER_VOID(LASER_SCAN_RECEIVED));
+
+  LaserScan laser_scan_fake(PSENscanInternalAngle(1), PSENscanInternalAngle(1), PSENscanInternalAngle(2));
   laser_scan_fake.measures_.push_back(1);
+  std::unique_ptr<MockScanner> mock_scanner{ new MockScanner() };
+  EXPECT_CALL(*(mock_scanner), getCompleteScan()).WillRepeatedly(Return(laser_scan_fake));
+  ROSScannerNode ros_scanner_node(nh_priv_, "scan", "scanner", Degree(137.5), std::move(mock_scanner));
 
-  LaserScan laser_scan_error_1{ PSENscanInternalAngle(1), PSENscanInternalAngle(1), PSENscanInternalAngle(2) };
-  laser_scan_error_1.measures_.push_back(1);
-  laser_scan_error_1.measures_.push_back(2);
+  subscriber.initialize(nh_priv_);
+  std::future<void> loop = std::async(std::launch::async, [&ros_scanner_node]() { ros_scanner_node.processingLoop(); });
+  BARRIER(LASER_SCAN_RECEIVED);
+  ros_scanner_node.terminateProcessingLoop();
+  EXPECT_EQ(loop.wait_for(LOOP_END_TIMEOUT), std::future_status::ready);
+}
 
-  LaserScan laser_scan_error_2{ PSENscanInternalAngle(0), PSENscanInternalAngle(1), PSENscanInternalAngle(2) };
-  laser_scan_error_2.measures_.push_back(1);
+ACTION(ThrowScanBuildFailure)
+{
+  throw LaserScanBuildFailure();
+}
 
-  LaserScan laser_scan_error_3{ PSENscanInternalAngle(10), PSENscanInternalAngle(2), PSENscanInternalAngle(1) };
-  laser_scan_error_3.measures_.push_back(1);
+TEST_F(RosScannerNodeTests, testScanBuildFailure)
+{
+  SubscriberMock subscriber;
+  EXPECT_CALL(subscriber, callback(::testing::_)).Times(1).WillOnce(ACTION_OPEN_BARRIER_VOID(LASER_SCAN_RECEIVED));
 
-  ros_scanner_node.buildRosMessage(laser_scan_fake);
-  EXPECT_THROW(ros_scanner_node.buildRosMessage(laser_scan_error_1), BuildROSMessageException);
-  EXPECT_THROW(ros_scanner_node.buildRosMessage(laser_scan_error_2), BuildROSMessageException);
-  EXPECT_THROW(ros_scanner_node.buildRosMessage(laser_scan_error_3), BuildROSMessageException);
+  LaserScan laser_scan_fake(PSENscanInternalAngle(1), PSENscanInternalAngle(1), PSENscanInternalAngle(2));
+  laser_scan_fake.measures_.push_back(1);
+  std::unique_ptr<MockScanner> mock_scanner{ new MockScanner() };
+  {
+    ::testing::InSequence s;
+    EXPECT_CALL(*(mock_scanner), getCompleteScan())
+        .Times(100)
+        .WillRepeatedly(DoAll(ThrowScanBuildFailure(), Return(laser_scan_fake)));
+    EXPECT_CALL(*(mock_scanner), getCompleteScan()).Times(1).WillRepeatedly(Return(laser_scan_fake));
+  }
+  ROSScannerNode ros_scanner_node(nh_priv_, "scan", "scanner", Degree(137.5), std::move(mock_scanner));
+
+  subscriber.initialize(nh_priv_);
+  std::future<void> loop = std::async(std::launch::async, [&ros_scanner_node]() { ros_scanner_node.processingLoop(); });
+  BARRIER(LASER_SCAN_RECEIVED);
+  ros_scanner_node.terminateProcessingLoop();
+  EXPECT_EQ(loop.wait_for(LOOP_END_TIMEOUT), std::future_status::ready);
+}
+
+TEST_F(RosScannerNodeTests, testMissingScannerObject)
+{
+  EXPECT_THROW(ROSScannerNode scanner_node(nh_priv_,
+                                           "scan",
+                                           "scanner",
+                                           Degree(137.5),
+                                           nullptr  // This throws exception
+                                           ),
+               psen_scan::PSENScanFatalException);
 }
 
 }  // namespace psen_scan_test
 
-int main(int argc, char** argv)
+int main(int argc, char* argv[])
 {
   ros::init(argc, argv, "integrationtest_ros_scanner_node");
   ros::NodeHandle nh;
