+++ test/integration_tests/integrationtest_ros_scanner_node.test
@@ -1,5 +1,5 @@
 <!--
-Copyright (c) 2019 Pilz GmbH & Co. KG
+Copyright (c) 2020 Pilz GmbH & Co. KG
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU Lesser General Public License as published by
@@ -14,8 +14,8 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 -->
-
 <launch>
-  <test pkg="psen_scan" test-name="integrationtest_ros_scanner_node" type="integrationtest_ros_scanner_node">
-  </test>
+
+  <test test-name="integrationtest_ros_scanner_node" pkg="psen_scan" type="integrationtest_ros_scanner_node"/>
+
 </launch>
